math/test: fix type mismatch warning on clang-20.
Fixes warning on clang-20 due to type mismatch.

+++ math/test/c23_references.h
@@ -199,6 +199,6 @@ arm_math_log2p1 (double x)
 long double
 arm_math_log2p1l (long double x)
 {
-  return (fabs (x) < 0x1p-52) ? (long double) x * M_INV_LOG2l
-			      : (log1p ((long double) x) / log (2));
+  return (fabsl (x) < 0x1p-52l) ? (long double) x * M_INV_LOG2l
+				: (log1pl ((long double) x) / logl (2));
 }
